Split LSP document-sync handlers into server_documents.go
Move the textDocument didOpen/didChange/didSave/didClose handlers, the
workspace didChangeWatchedFiles/didChangeConfiguration events, and the
session-overlay helpers (syncBuffer, dropPath, openDocPaths,
watchedFilesTreeChanged) out of server.go. Pure file-split refactor —
no behavior change. server.go: 2008 -> 1756 lines.
